refactor: consolidate controller rendering
